# docs/swagger.yaml
basePath: /userApp
definitions:
  swagger.AplicationUser:
    properties:
      userNameapp:
        type: string
    type: object
  swagger.LoginForm:
    properties:
      email:
        type: string
      password:
        type: string
      userNameApp:
        type: string
    type: object
  views.Response:
    properties:
      accessToken:
        type: string
      failedAttempts:
        type: integer
      message:
        type: string
      refreshToken:
        type: string
      success:
        type: boolean
      users:
        $ref: '#/definitions/views.Users'
    type: object
  views.Users:
    properties:
      email:
        type: string
      fkCompany:
        type: integer
      fkCustomer:
        type: integer
      fkRole:
        type: integer
      id:
        type: integer
      lastLogin:
        type: string
      lastname:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
  version: "1"
paths:
  /login/userApp/:
    post:
      consumes:
      - application/json
      description: Autentica a un usuario y devuelve un token de acceso y los detalles
        del usuario.
      parameters:
      - description: Credenciales del usuario
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/swagger.LoginForm'
      produces:
      - application/json
      responses:
        "200":
          description: Respuesta exitosa con token y detalles del usuario
          schema:
            $ref: '#/definitions/views.Response'
        "400":
          description: 'Error: Datos inválidos o problema con el formato del email'
        "401":
          description: 'Error: Credenciales incorrectas o intento de inicio de sesión
            fallido'
        "500":
          description: Error interno del servidor
      summary: Iniciar sesión
      tags:
      - appUser
  /userApp/:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        description: Número de página para la paginación
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties:
                type: string
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Obtiene todos Ususarios de palicacion
      tags:
      - appUser
  /userApp/{id}:
    put:
      consumes:
      - application/json
      description: Actualiza un usuario de la aplicación
      parameters:
      - description: ID del usuario de la aplicación
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Actualizar un userApp
      tags:
      - appUser
  /userApp/create/:
    post:
      consumes:
      - application/json
      description: Añade un nuevo usuario de aplicación a la base de datos
      parameters:
      - description: AplicationUser
        in: body
        name: vehiculo
        required: true
        schema:
          $ref: '#/definitions/swagger.AplicationUser'
      produces:
      - application/json
      responses:
        "200":
          description: ID o mensaje de éxito del AplicationUser creado
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Crea un nuevo usuario de aplicación
      tags:
      - appUser
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"